+++ core/src/main/java/com/cartoonsurvivors/game/utilidades/EstadoJuego.java
@@ -1,0 +1,7 @@
+package com.cartoonsurvivors.game.utilidades;
+
+public enum EstadoJuego {
+    JUGANDO,
+    PAUSADO,
+    GAME_OVER
+}