+++ CHANGELOG.md
@@ -10,6 +10,16 @@ Todos los cambios importantes del proyecto se documentan en este archivo.
 
 ---
 
+## [0.1.19] 2026-09-04
+
+### Agregado
+
+* Se agrego la clase PantallaPausa con funciones de pausa, reanudación del juego y salida al menú principal.
+* Se agrego la funcion de pausa y play en la clase de AudioManager para poder pausar y reanudar la musica del juego.
+* Se agrego un enum con los estados del juego para poder controlar el flujo del mismo.
+* Se agrego una funcion para poder detectar el escape en controlador de entrada y pausar el juego.
+---
+
 ## [0.1.17] 2026-09-03
 
 ### Agregado